% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A positive integer}
}
\value{
A character vector
}
\description{
The goal is to count as high as you can, but for any number evenly
divisible by 3, substitute "Fizz" and any number evenly divisible by 5,
substitute "Buzz", and if it is divisible by both, substitute "Fizz Buzz".
}
\examples{
FizzBuzz(50)
FizzBuzz(100)
}
