% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.}
\format{
a data frame with 22967 observations
\describe{
  \item{DATE}{Date in month-day-year format.}
  \item{PRCP}{Percipitation for a given date in inches.}
  \item{SNOW}{Amount of snow in inches for a given day.}
  \item{TMAX}{Max temperature in degrees Farenheight for a given day.}
  \item{TMIN}{Min temperature in degrees Farenheight for a given day.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
Weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.
}
\keyword{datasets}
